+++ rays.go
@@ -17,7 +17,7 @@ func (a *Ray) String() string {
 	return fmt.Sprintf("origin=%v, direction=%v", a.origin, a.direction)
 }
 
-func (a *Ray) Position(t float32) *Tuple {
+func (a *Ray) Position(t float64) *Tuple {
 	return a.direction.Multiply(t).Add(a.origin)
 }
