+++ tuples.go
@@ -9,8 +9,8 @@ import (
 const Epsilon = 0.00001
 
 type Tuple struct {
-	X, Y, Z float32 // 3D coordinates
-	W       float32 // 1.0 when a point, 0.0 when a vector
+	X, Y, Z float64 // 3D coordinates
+	W       float64 // 1.0 when a point, 0.0 when a vector
 }
 
 // String returns a string representation of the tuple
@@ -70,7 +70,7 @@ func (a *Tuple) Negate() *Tuple {
 }
 
 // Multiply multiplies a tuple by a scalar
-func (a *Tuple) Multiply(scalar float32) *Tuple {
+func (a *Tuple) Multiply(scalar float64) *Tuple {
 	return &Tuple{
 		X: a.X * scalar,
 		Y: a.Y * scalar,
@@ -80,7 +80,7 @@ func (a *Tuple) Multiply(scalar float32) *Tuple {
 }
 
 // Divide divides a tuple by a scalar
-func (a *Tuple) Divide(scalar float32) *Tuple {
+func (a *Tuple) Divide(scalar float64) *Tuple {
 	return &Tuple{
 		X: a.X / scalar,
 		Y: a.Y / scalar,
@@ -90,12 +90,12 @@ func (a *Tuple) Divide(scalar float32) *Tuple {
 }
 
 // Magnitude returns the magnitude (or length) of the tuple
-func (a *Tuple) Magnitude() float32 {
+func (a *Tuple) Magnitude() float64 {
 	x := (float64)(a.X)
 	y := (float64)(a.Y)
 	z := (float64)(a.Z)
 	w := (float64)(a.W)
-	return (float32)(math.Sqrt(x*x + y*y + z*z + w*w))
+	return (float64)(math.Sqrt(x*x + y*y + z*z + w*w))
 }
 
 // Normalize returns a normalized unit vector
@@ -110,7 +110,7 @@ func (a *Tuple) Normalize() *Tuple {
 }
 
 // Dot returns the dot product of this vector with the provided vector
-func (a *Tuple) Dot(b *Tuple) float32 {
+func (a *Tuple) Dot(b *Tuple) float64 {
 	return a.X*b.X + a.Y*b.Y + a.Z*b.Z + a.W*b.W
 }
 
@@ -123,13 +123,13 @@ func (a *Tuple) Cross(b *Tuple) *Tuple {
 	)
 }
 
-// EqualFloat determines if two float32 values are the within Epsilon of each other
-func EqualFloat32(a, b float32) bool {
+// EqualFloat determines if two float64 values are the within Epsilon of each other
+func EqualFloat32(a, b float64) bool {
 	return math.Abs((float64)(a)-(float64)(b)) < Epsilon
 }
 
 // NewPoint creates a new Tuple of type point
-func NewPoint(X, Y, Z float32) *Tuple {
+func NewPoint(X, Y, Z float64) *Tuple {
 	point := &Tuple{
 		X: X,
 		Y: Y,
@@ -141,7 +141,7 @@ func NewPoint(X, Y, Z float32) *Tuple {
 }
 
 // NewVector creates a new Tuple of type vector
-func NewVector(X, Y, Z float32) *Tuple {
+func NewVector(X, Y, Z float64) *Tuple {
 	vector := &Tuple{
 		X: X,
 		Y: Y,
@@ -153,7 +153,7 @@ func NewVector(X, Y, Z float32) *Tuple {
 }
 
 type Color struct {
-	Red, Green, Blue float32
+	Red, Green, Blue float64
 }
 
 // Equal determines if two Colors are the same
@@ -189,7 +189,7 @@ func (a *Color) Multiply(b *Color) *Color {
 }
 
 // MultiplyScalar multiplies this Color by a scalar
-func (a *Color) MultiplyScalar(scalar float32) *Color {
+func (a *Color) MultiplyScalar(scalar float64) *Color {
 	return &Color{
 		Red:   a.Red * scalar,
 		Green: a.Green * scalar,
